refactor: 멤버 인수 테스트 리팩토링
